Extract document metadata package
