+++ src/main/java/com/code4ro/legalconsultation/config/mail/MailConfiguration.java
@@ -1,6 +1,6 @@
 package com.code4ro.legalconsultation.config.mail;
 
-import com.code4ro.legalconsultation.model.persistence.DocumentMetadata;
+import com.code4ro.legalconsultation.document.metadata.model.persistence.DocumentMetadata;
 import com.code4ro.legalconsultation.invitation.model.persistence.Invitation;
 import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.service.api.MailApi;
