+++ src/main/java/com/code4ro/legalconsultation/document/metadata/model/persistence/DocumentMetadata.java
@@ -1,5 +1,7 @@
-package com.code4ro.legalconsultation.model.persistence;
+package com.code4ro.legalconsultation.document.metadata.model.persistence;
 
+import com.code4ro.legalconsultation.model.persistence.BaseEntity;
+import com.code4ro.legalconsultation.model.persistence.DocumentType;
 import lombok.Getter;
 import lombok.Setter;
