+++ src/main/java/com/code4ro/legalconsultation/model/dto/DocumentConsolidatedDto.java
@@ -1,5 +1,6 @@
 package com.code4ro.legalconsultation.model.dto;
 
+import com.code4ro.legalconsultation.document.metadata.model.dto.DocumentMetadataDto;
 import com.code4ro.legalconsultation.document.node.model.dto.DocumentNodeDto;
 import com.code4ro.legalconsultation.user.model.dto.UserDto;
 import lombok.Getter;
