+++ src/test/java/com/code4ro/legalconsultation/controller/DocumentControllerIntegrationTest.java
@@ -9,10 +9,10 @@ import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import com.code4ro.legalconsultation.model.dto.DocumentViewDto;
 import com.code4ro.legalconsultation.model.persistence.DocumentConfiguration;
 import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
-import com.code4ro.legalconsultation.model.persistence.DocumentMetadata;
+import com.code4ro.legalconsultation.document.metadata.model.persistence.DocumentMetadata;
 import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
 import com.code4ro.legalconsultation.repository.DocumentConsolidatedRepository;
-import com.code4ro.legalconsultation.repository.DocumentMetadataRepository;
+import com.code4ro.legalconsultation.document.metadata.repository.DocumentMetadataRepository;
 import com.code4ro.legalconsultation.document.node.repository.DocumentNodeRepository;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
